requirements.txt
entsoe-py ==0.2.10
pandas
python-dotenv
pycountry
